Update cli_admin/, other... Добавлена команда find в cli_admin, а также исправлены ошибки с работой с ингредиентами.

+++ server/app/admin_cli/commands/delete.py
@@ -11,6 +11,22 @@ class Delete(Base):
             print('Указано неверное имя таблицы.')
             return
 
+        if id == -1:
+            answer = input(f'Вы точно хотите удалить ВСЕ записи таблицы {tablename}? (Y/n): ')
+            if answer != 'Y':
+                print('Прерываю удаление.')
+                return
+            i = 1
+            while True:
+                try:
+                    record = repository.get(i)
+                    repository.delete(record)
+                except Exception:
+                    break
+                i += 1
+            print('Все записи успешно удалены.')
+            return
+
         try:
             record = repository.get(id)
         except Exception as e:
@@ -23,7 +39,13 @@ class Delete(Base):
             return
 
         try:
-            repository.delete(record)
+            answer = input(f'Вы точно хотите удалить запись из '
+                           f'таблицы {tablename} с индексом {id}? (Y/n): ')
+            if answer == 'Y':
+                repository.delete(record)
+            else:
+                print('Прерываю удаление')
+                return
         except Exception as e:
             print('При удалении возникли ошибки.')
             print(e)
